playwright/requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
joblib==1.4.2
numpy==1.26.4
packaging==24.1
pandas==2.2.2
playwright==1.44.0
pluggy==1.5.0
pyee==11.1.0
pytest==8.2.2
pytest-base-url==2.1.0
pytest-playwright==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
soupsieve==2.5
text-unidecode==1.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
